requirements.txt
# conda install pytorch torchvision torchaudio cudatoolkit=11.3 -c pytorch
numpy
PyYAML
imageio
imageio-ffmpeg
matplotlib
opencv-python
scikit-image
tqdm
h5py
progressbar
psutil
torch
torchaudio
torchvision
ninja
pillow
torchmetrics
